backend/requirements.txt
Django
psycopg2
python-decouple
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers
Pillow
PyJWT
django-import-export
django-filter
passlib
cryptography
xhtml2pdf
crispy-bootstrap5